+++ Wallet-System/User.h
@@ -15,7 +15,7 @@ public:
   string getUsername();
   string getPassword();
   string getTotpSecret();
-  float getBalance();
+  double getBalance();
   bool getSuspendedFlag();
   bool getIsHas2FA();
 
@@ -23,14 +23,14 @@ public:
   void setIsHas2FA(bool flag);
   bool setUsername(string newUsername);
   void setPassword(string newPassword);
-  void setBalance(float amount);
+  void setBalance(double amount);
   void setSuspendedFlag(bool flag);
 
   bool isUniqueUsername();
-  void addMoney(float value);
+  void addMoney(double value);
   bool strongPassword(string password);
   bool isSentTransaction(Transaction *transaction);
-  void requestMoney(string username, float amount);
+  void requestMoney(string username, double amount);
 
   vector<Transaction*> getAllTransactions();
   vector<Transaction*> getToRequests();
@@ -38,8 +38,8 @@ public:
 
   vector<Transaction*> getReceivedTransactions();
   vector<Transaction*> getSentTransactions();
-  void makeTransaction(string receiver, float amount);
-  void makeRequest(string requested, float amount);
+  void makeTransaction(string receiver, double amount);
+  void makeRequest(string requested, double amount);
 
   friend std::ostream& operator<<(std::ostream& os, User& t);
   static bool compareByUsername(const User* a, const User* b);
@@ -50,7 +50,7 @@ protected:
 private:
   string totpSecret;
   bool isHas2FA = false;
-  float balancee;
+  double balancee;
   bool suspendedFlag;
 };
 
@@ -58,24 +58,24 @@ class Transaction {
 private:
   string sender;
   string recipient;
-  float amountt;
+  double amountt;
   int isPending;
   chrono::system_clock::time_point dateTime;
 
 public:
   bool operator>(const Transaction& other) const;
   Transaction();
-  Transaction(string sender, string recipient, float amount);
+  Transaction(string sender, string recipient, double amount);
   string getSenderUserName();
   string getRecipientUserName();
-  float getAmount();
+  double getAmount();
   int getIsPending();
   chrono::system_clock::time_point getDateTime();
   void setDateTime(chrono::system_clock::time_point date);
   void setIsPending(int pending);
   void setSenderUsername(string username);
   void setRecipientUsername(string username);
-  void setAmount(float amount);
+  void setAmount(double amount);
 
   void declineTransaction();
   void acceptTransaction();
